Make it possible to use remote custom registries
This is accomplished by adding a generic abstraction over local and
remote resources in `pkg/fetch` with the `fetch.Fetch` and
`fetch.ToFile` functions.

This generic abstraction is meant to _gradually_ replace the download
infrastructure used in the rest of just-install and it's currently used
only to fetch custom registries.

This commit also introduces the `pkg/paths` package that will gradually
take over similar functions used in `pkg/justinstall`.